refactor(runners): extract build_superstep_records to shared helper
Move StepRecord-building logic from AsyncRunner._save_superstep_records
into runners/_shared/checkpoint_helpers.py as a pure function. Both
sync and async runners will call this shared helper. AsyncRunner now
delegates to it and only handles durability dispatch.

Also adds _compute_child_run_id for deterministic GraphNode child IDs,
preparing for hierarchical checkpointing.
